Basic parsing, interpretation, and addition
This commit adds the LICENSE and README.md files, but more importantly,
it also adds the first, most basic interpreter that we'll create.

Since we're starting from scratch, and not building on other commits
the way that later versions will, it's a big chunk of code to digest.
If you can understand what's going on here, you'll have no trouble
following the rest of the code that we develop. If you are a bit
confused about what exactly the code in this commit does, there
are a few  things you should do:

1) Read other resources to explain any terms you don't understand
2) Play with the code, to see how it breaks when you change stuff
3) Open an issue with your question on the GitHub repository: [1]

[1]: https://github.com/setupminimal/myoplitmol