Defining words
In this commit, we let users of our programming language define their
own words. With this, the language is almost usable - you can have
code reuse, and all that jazz, but it's not quite Turing-complete.

For a language to be Turing-complete (that is, capable of anything that
a normal programming language would be capable of), it must be able to
loop. With the ability to define our own words, we can loop (Try `:
infinite-loop infinite-loop ;`), but we can't _stop_ looping. In the
next commit, we add `if`s to the language, which let us have finite loops.