Adding Loops
In this commit, we implement loops, the final feature needed to be
Turing-complete. In our case, loops and if statements all come from
the same feature, so adding them is pretty straightforward.

At this point, the language is complete. There's one more commit that
cleans up the command line interface/modularizes the code a little bit,
but at this point, you should be thinking of what the takeaway is here:

- Making a minimal language is really easy
- Even though it's technically as powerful as other programming
  languages, it's still pretty miserable to use
- Adding very powerful features, like loops, can be done with very
  little code, if you think about it a bit

At this point, I strongly recommend playing around with the code, trying
to write programs in our language, and adding new features. You might
try adding imports, variables, infix operators, or anything else you can
think of. Don't be scared to make mistakes - that's what version control
is for.

When you eventually run up against the limitations of what we've done
here, you might also check out the book Crafting
Interpreters (craftinginterpreters.com), which is freely available
online. It does a great job of building a more traditional language with
braces, functions, loops, bytecode, etc. It also goes into more depth on
topics like Parsing that we completely skipped in this language for the
sake of time.

And finally, I hope that you use this experience as inspiration to try
and build something that's uniquely yours. Making programming languages
is fascinating and rewarding - I hope you can enjoy it as much as I do.